x64 warning C4244: '=' : conversion from 'SOCKET' to 'int', possible loss of data

+++ lib/pollfd.c
@@ -32,9 +32,9 @@ insert_wsi_socket_into_fds(struct libwebsocket_context *context,
 		return 1;
 	}
 
-	if (wsi->sock >= context->max_fds) {
+	if (wsi->sock >= (unsigned)context->max_fds) {
 		lwsl_err("Socket fd %d is too high (%d)\n",
-						wsi->sock, context->max_fds);
+						(int)wsi->sock, context->max_fds);
 		return 1;
 	}
 
@@ -83,7 +83,7 @@ remove_wsi_socket_from_fds(struct libwebsocket_context *context,
 		goto do_ext;
 	}
 
-	if (wsi->sock > context->max_fds) {
+	if (wsi->sock > (unsigned)context->max_fds) {
 		lwsl_err("Socket fd %d too high (%d)\n",
 						   wsi->sock, context->max_fds);
 		return 1;
